requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
django-ckeditor>=6.7.0
python-dotenv>=1.0.0
